Rename template component to template container
