+++ src/app/feature/template-testing/template-testing.page.html
@@ -1,3 +1,3 @@
 <ion-content>
-  <plh-template *ngIf="testTemplate" [template]="testTemplate"></plh-template>
+  <plh-template-container *ngIf="testTemplate" [template]="testTemplate"></plh-template-container>
 </ion-content>
