+++ src/app/shared/components/template/components/display_group.ts
@@ -1,12 +1,12 @@
 import { Component, Input, OnInit } from "@angular/core";
 import { BehaviorSubject } from "rxjs";
 import { FlowTypes } from 'src/app/shared/model/flowTypes';
-import { ITemplateComponent } from "./tmpl-comp-host";
+import { ITemplateComponent } from "./tmpl.component";
 
 @Component({
   selector: "plh-tmpl-display-group",
   template: `<div class="display-group">
-    <plh-tmpl-comp-host *ngFor="let childRow of row.rows" [row]="childRow" [template]="template" [localVariables]="localVariables"></plh-tmpl-comp-host>
+    <plh-tmpl-comp *ngFor="let childRow of row.rows" [row]="childRow" [template]="template" [localVariables]="localVariables"></plh-tmpl-comp>
   </div>`,
   styleUrls: ["./tmpl-components-common.scss"]
 })
