+++ src/app/shared/components/template/template-container.component.html
@@ -1,6 +1,6 @@
 <div [className]="'template-container ' + localVariables?.theme_name">
   <div *ngFor="let row of _template?.rows" class="row">
-    <plh-tmpl-comp-host [row]="row" [template]="_template" [localVariables]="localVariables"></plh-tmpl-comp-host>
+    <plh-tmpl-comp [row]="row" [template]="_template" [localVariables]="localVariables"></plh-tmpl-comp>
   </div>
   <h4>Local Variables</h4>
   <ul>
