+++ src/app/shared/components/template/template-container.component.scss
