+++ src/app/shared/components/template/template-container.component.ts
@@ -1,14 +1,12 @@
-import { ChangeDetectorRef, Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
-import { BehaviorSubject } from 'scripts/node_modules/rxjs';
+import { ChangeDetectorRef, Component, Input } from '@angular/core';
 import { FlowTypes } from '../../model';
-import { TEMPLATE } from '../../services/data/data.service';
 
 @Component({
-  selector: 'plh-template',
-  templateUrl: './template.component.html',
-  styleUrls: ['./template.component.scss'],
+  selector: 'plh-template-container',
+  templateUrl: './template-container.component.html',
+  styleUrls: ['./template-container.component.scss'],
 })
-export class TemplateComponent {
+export class TemplateContainerComponent {
 
   _template: FlowTypes.Template;
   @Input() set template(value: FlowTypes.Template) {
